Eliminate dependecy `tensorflow.model`
