--- unsupervised/translate/__init__.py
@@ -1,4 +1,0 @@
-"""Seq2seq translate model from tensorflow tutorial.
-
-Copied from https://github.com/tensorflow/models/blob/master/tutorials/rnn/translate/.
-"""
